Add base class for youtube stream extractor tests
