+++ extractor/src/test/java/org/schabi/newpipe/extractor/services/youtube/stream/YoutubeStreamExtractorBaseTest.java
@@ -1,0 +1,114 @@
+package org.schabi.newpipe.extractor.services.youtube.stream;
+
+import org.junit.Test;
+import org.schabi.newpipe.extractor.exceptions.ExtractionException;
+import org.schabi.newpipe.extractor.exceptions.ParsingException;
+import org.schabi.newpipe.extractor.services.youtube.extractors.YoutubeStreamExtractor;
+import org.schabi.newpipe.extractor.stream.AudioStream;
+import org.schabi.newpipe.extractor.stream.StreamInfoItemsCollector;
+import org.schabi.newpipe.extractor.stream.StreamType;
+import org.schabi.newpipe.extractor.stream.VideoStream;
+import org.schabi.newpipe.extractor.utils.Utils;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.junit.Assert.*;
+import static org.schabi.newpipe.extractor.ExtractorAsserts.*;
+
+public abstract class YoutubeStreamExtractorBaseTest {
+    public static YoutubeStreamExtractor extractor;
+
+    @Test
+    public void testGetAgeLimit() throws ParsingException {
+        assertEquals(YoutubeStreamExtractor.NO_AGE_LIMIT, extractor.getAgeLimit());
+    }
+
+    @Test
+    public void testGetTitle() throws ParsingException {
+        assertFalse(extractor.getName().isEmpty());
+    }
+
+    @Test
+    public void testGetDescription() throws ParsingException {
+        assertNotNull(extractor.getDescription());
+        assertFalse(extractor.getDescription().isEmpty());
+    }
+
+    @Test
+    public void testGetUploaderName() throws ParsingException {
+        assertNotNull(extractor.getUploaderName());
+        assertFalse(extractor.getUploaderName().isEmpty());
+    }
+
+    @Test
+    public void testGetLength() throws ParsingException {
+        long length = extractor.getLength();
+        assertTrue(String.valueOf(length), length > 0);
+    }
+
+    @Test
+    public void testGetViewCount() throws ParsingException {
+        long count = extractor.getViewCount();
+        assertTrue(Long.toString(count), count > 0);
+    }
+
+    @Test
+    public void testGetUploadDate() throws ParsingException {
+        assertTrue(extractor.getUploadDate().length() > 0);
+    }
+
+    @Test
+    public void testGetUploaderUrl() throws ParsingException {
+        String uploaderUrl = extractor.getUploaderUrl();
+        assertTrue(uploaderUrl.startsWith("https://www.youtube.com/channel/"));
+    }
+
+    @Test
+    public void testGetThumbnailUrl() throws ParsingException {
+        assertIsSecureUrl(extractor.getThumbnailUrl());
+    }
+
+    @Test
+    public void testGetUploaderAvatarUrl() throws ParsingException {
+        assertIsSecureUrl(extractor.getUploaderAvatarUrl());
+    }
+
+    @Test
+    public void testGetAudioStreams() throws IOException, ExtractionException {
+        for (AudioStream s : extractor.getAudioStreams()) {
+            assertIsSecureUrl(s.url);
+            assertResponseCodeOk(s.url);
+        }
+    }
+
+    @Test
+    public void testGetVideoStreams() throws IOException, ExtractionException {
+        List<VideoStream> streams = new ArrayList<>();
+        streams.addAll(extractor.getVideoStreams());
+        streams.addAll(extractor.getVideoOnlyStreams());
+
+        for (VideoStream s : streams) {
+            assertIsSecureUrl(s.url);
+            assertResponseCodeOk(s.url);
+            assertTrue(s.resolution.length() > 0);
+            assertTrue(Integer.toString(s.getFormatId()),
+                    0 <= s.getFormatId() && s.getFormatId() <= 0x100);
+        }
+        assertFalse(streams.isEmpty());
+    }
+
+    @Test
+    public void testStreamType() throws ParsingException {
+        assertTrue(extractor.getStreamType() == StreamType.VIDEO_STREAM);
+    }
+
+    @Test
+    public void testGetRelatedVideos() throws ExtractionException, IOException {
+        StreamInfoItemsCollector relatedVideos = extractor.getRelatedStreams();
+        Utils.printErrors(relatedVideos.getErrors());
+        assertFalse(relatedVideos.getItems().isEmpty());
+        assertTrue(relatedVideos.getErrors().isEmpty());
+    }
+}
