+++ extractor/src/test/java/org/schabi/newpipe/extractor/services/youtube/stream/YoutubeStreamExtractorDefaultTest.java
@@ -11,13 +11,11 @@ import org.schabi.newpipe.extractor.exceptions.ParsingException;
 import org.schabi.newpipe.extractor.services.youtube.extractors.YoutubeStreamExtractor;
 import org.schabi.newpipe.extractor.stream.*;
 import org.schabi.newpipe.extractor.utils.Localization;
-import org.schabi.newpipe.extractor.utils.Utils;
 
 import java.io.IOException;
 import java.util.List;
 
 import static org.junit.Assert.*;
-import static org.schabi.newpipe.extractor.ExtractorAsserts.assertIsSecureUrl;
 import static org.schabi.newpipe.extractor.ServiceList.YouTube;
 
 /*
@@ -48,9 +46,7 @@ public class YoutubeStreamExtractorDefaultTest {
     /**
      * Test for {@link StreamExtractor}
      */
-    public static class AdeleHello {
-        private static YoutubeStreamExtractor extractor;
-
+    public static class AdeleHello extends YoutubeStreamExtractorBaseTest {
         @BeforeClass
         public static void setUp() throws Exception {
             NewPipe.init(Downloader.getInstance(), new Localization("GB", "en"));
@@ -72,78 +68,22 @@ public class YoutubeStreamExtractorDefaultTest {
         }
 
         @Test
-        public void testGetTitle() throws ParsingException {
-            assertFalse(extractor.getName().isEmpty());
-        }
-
-        @Test
-        public void testGetDescription() throws ParsingException {
-            assertNotNull(extractor.getDescription());
-            assertFalse(extractor.getDescription().isEmpty());
-        }
-
-        @Test
         public void testGetFullLinksInDescription() throws ParsingException {
             assertTrue(extractor.getDescription().contains("http://adele.com"));
             assertFalse(extractor.getDescription().contains("http://smarturl.it/SubscribeAdele?IQi..."));
         }
 
         @Test
-        public void testGetUploaderName() throws ParsingException {
-            assertNotNull(extractor.getUploaderName());
-            assertFalse(extractor.getUploaderName().isEmpty());
-        }
-
-
-        @Test
+        @Override
         public void testGetLength() throws ParsingException {
             assertEquals(366, extractor.getLength());
         }
 
         @Test
+        @Override
         public void testGetViewCount() throws ParsingException {
             Long count = extractor.getViewCount();
             assertTrue(Long.toString(count), count >= /* specific to that video */ 1220025784);
-        }
-
-        @Test
-        public void testGetUploadDate() throws ParsingException {
-            assertTrue(extractor.getUploadDate().length() > 0);
-        }
-
-        @Test
-        public void testGetUploaderUrl() throws ParsingException {
-            assertEquals("https://www.youtube.com/channel/UCsRM0YB_dabtEPGPTKo-gcw", extractor.getUploaderUrl());
-        }
-
-        @Test
-        public void testGetThumbnailUrl() throws ParsingException {
-            assertIsSecureUrl(extractor.getThumbnailUrl());
-        }
-
-        @Test
-        public void testGetUploaderAvatarUrl() throws ParsingException {
-            assertIsSecureUrl(extractor.getUploaderAvatarUrl());
-        }
-
-        @Test
-        public void testGetAudioStreams() throws IOException, ExtractionException {
-            assertFalse(extractor.getAudioStreams().isEmpty());
-        }
-
-        @Test
-        public void testGetVideoStreams() throws IOException, ExtractionException {
-            for (VideoStream s : extractor.getVideoStreams()) {
-                assertIsSecureUrl(s.url);
-                assertTrue(s.resolution.length() > 0);
-                assertTrue(Integer.toString(s.getFormatId()),
-                        0 <= s.getFormatId() && s.getFormatId() <= 0x100);
-            }
-        }
-
-        @Test
-        public void testStreamType() throws ParsingException {
-            assertTrue(extractor.getStreamType() == StreamType.VIDEO_STREAM);
         }
 
         @Test
@@ -151,14 +91,6 @@ public class YoutubeStreamExtractorDefaultTest {
             // we dont expect this particular video to have a DASH file. For this purpouse we use a different test class.
             assertTrue(extractor.getDashMpdUrl(),
                     extractor.getDashMpdUrl() != null && extractor.getDashMpdUrl().isEmpty());
-        }
-
-        @Test
-        public void testGetRelatedVideos() throws ExtractionException, IOException {
-            StreamInfoItemsCollector relatedVideos = extractor.getRelatedStreams();
-            Utils.printErrors(relatedVideos.getErrors());
-            assertFalse(relatedVideos.getItems().isEmpty());
-            assertTrue(relatedVideos.getErrors().isEmpty());
         }
 
         @Test
@@ -174,21 +106,13 @@ public class YoutubeStreamExtractorDefaultTest {
         }
     }
 
-    public static class DescriptionTestPewdiepie {
-        private static YoutubeStreamExtractor extractor;
-
+    public static class DescriptionTestPewdiepie extends YoutubeStreamExtractorBaseTest {
         @BeforeClass
         public static void setUp() throws Exception {
             NewPipe.init(Downloader.getInstance(), new Localization("GB", "en"));
             extractor = (YoutubeStreamExtractor) YouTube
                     .getStreamExtractor("https://www.youtube.com/watch?v=fBc4Q_htqPg");
             extractor.fetchPage();
-        }
-
-        @Test
-        public void testGetDescription() throws ParsingException {
-            assertNotNull(extractor.getDescription());
-            assertFalse(extractor.getDescription().isEmpty());
         }
 
         @Test
@@ -203,21 +127,13 @@ public class YoutubeStreamExtractorDefaultTest {
         }
     }
 
-    public static class DescriptionTestUnboxing {
-        private static YoutubeStreamExtractor extractor;
-
+    public static class DescriptionTestUnboxing extends YoutubeStreamExtractorBaseTest {
         @BeforeClass
         public static void setUp() throws Exception {
             NewPipe.init(Downloader.getInstance(), new Localization("GB", "en"));
             extractor = (YoutubeStreamExtractor) YouTube
                     .getStreamExtractor("https://www.youtube.com/watch?v=cV5TjZCJkuA");
             extractor.fetchPage();
-        }
-
-        @Test
-        public void testGetDescription() throws ParsingException {
-            assertNotNull(extractor.getDescription());
-            assertFalse(extractor.getDescription().isEmpty());
         }
 
         @Test
